+++ test/main/tcp.cpp
@@ -110,6 +110,9 @@ void TestTcp() {
 
   TEST_ASSERT(server.Disable() == ESP_OK);
   TEST_ASSERT(!server.IsEnabled());
+
+  TEST_ASSERT(ipV4Client.Disconnect() == ESP_OK);
+  TEST_ASSERT(ipV4Client.Connect() == ESP_FAIL);
 }
 
 //==============================================================================